requirements.txt
ephem>=4.1
numpy>=1.25
opencv-python-headless>=4.8
scipy>=1.11
Pillow>=10.0
pyyaml>=6.0
requests>=2.31
